+++ KmpCrypto/src/jvmTest/kotlin/com/oldguy/crypto/ModesTests.kt
@@ -1,8 +1,13 @@
 package com.oldguy.crypto
 
+import com.oldguy.common.io.ByteBuffer
+import com.oldguy.common.io.Charset
+import com.oldguy.common.io.Charsets
+import com.oldguy.common.io.UByteBuffer
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
+import kotlin.test.assertContentEquals
 
 @ExperimentalCoroutinesApi
 class AesTests {
@@ -12,17 +17,43 @@ class AesTests {
     @Test
     fun aesCbc() {
         runTest {
+            val payload = "any stuff in here 12345678"
+            val keyBytes = byteArrayOf(0,1,2,3,4,5,6,7,8,9,0x7f,0x55)
+            val payloadBytes = Charset(Charsets.Utf8).encode(payload)
             Cipher.build {
-                engine { aes() }
-                mode = CipherModes.CBC
                 padding = Paddings.PKCS7
+                engine { aes() }
                 key {
-                    stringKey = CryptoTestHelp.stringKey
-                    hashKeyLengthBits = 256
+                    key = keyBytes.toUByteArray()
                     keyDigest = Digests.SHA256
                 }
             }.apply {
-                CryptoTestHelp.singleBufferTest(this)
+                val encrypted = processOne(true, UByteBuffer(payloadBytes.toUByteArray()))
+
+                val javaCipher = org.bouncycastle.crypto.paddings.PaddedBufferedBlockCipher(
+//                    org.bouncycastle.crypto.modes.CBCBlockCipher(
+                    org.bouncycastle.crypto.engines.AESEngine()
+                    )
+//                ))
+                val hashKeyBytes = ByteArray(32)
+                org.bouncycastle.crypto.digests.SHA256Digest().apply {
+                    update(keyBytes, 0, keyBytes.size)
+                    doFinal(hashKeyBytes, 0)
+                }
+                val hashKey = org.bouncycastle.crypto.params.KeyParameter(hashKeyBytes)
+                javaCipher.init(true, hashKey)
+                val javaEncrypted =
+                    CryptoTestHelp.bouncyProcess(
+                        javaCipher,
+                        ByteBuffer(payloadBytes)
+                    )
+
+                val jb = javaEncrypted.getBytes()
+                val eb = encrypted.getBytes().toByteArray()
+                assertContentEquals(jb, eb)
+
+                val decrypted = processOne(false, encrypted.flip())
+                assertContentEquals(payloadBytes, decrypted.getBytes().toByteArray())
             }
         }
     }
@@ -34,7 +65,6 @@ class AesTests {
                 parse("aes/cfb")
                 key {
                     stringKey = CryptoTestHelp.stringKey
-                    hashKeyLengthBits = 256
                     keyDigest = Digests.SHA256
                 }
             }.apply {
@@ -52,7 +82,6 @@ class AesTests {
                     key = CryptoTestHelp.key1
                     ivSizeMatchBlock = false
                     iv = iv8
-                    hashKeyLengthBits = 256
                     keyDigest = Digests.SHA256
                 }
             }.apply {
@@ -69,7 +98,6 @@ class AesTests {
                 key {
                     key = CryptoTestHelp.key1
                     iv = randomIV()
-                    hashKeyLengthBits = 256
                     keyDigest = Digests.SHA256
                 }
             }.apply {
@@ -82,11 +110,10 @@ class AesTests {
     fun aesEcbTest() {
         runTest {
             Cipher.build {
-                parse("aes/ecb")
+                parse("aes/ecb/pkcs7")
                 key {
                     key = CryptoTestHelp.key1
                     iv = iv16
-                    hashKeyLengthBits = 256
                     keyDigest = Digests.SHA256
                 }
             }.apply {
@@ -99,13 +126,12 @@ class AesTests {
     fun aesCbcPKCSTest() {
         runTest {
             Cipher.build {
-                engine { aes() }
                 mode = CipherModes.CBC
                 padding = Paddings.PKCS7
+                engine { aes() }
                 key {
                     key = CryptoTestHelp.key1
-                    hashKeyLengthBits = 256
-                    keyDigest = Digests.SHA256
+                     keyDigest = Digests.SHA256
                 }
             }.apply {
                 CryptoTestHelp.singleBufferTest(this)